fix(s18-review): truncate AFTER multiply for fractional byte values
parseBytes was doing int64(f) * mult which truncated the float before
applying the unit multiplier. "1.5KB" returned 1000 (1*1000) instead of
1500. Fix: int64(f * float64(mult)) so the multiply happens in float64
before truncation.

Removes the t.Skipf on TestSpec_S21_5_FractionalByteValues; all 6 cases
(including new 2.5KiB→2560) now pass. Spec issue #74 resolved.

+++ config.go
@@ -486,7 +486,7 @@ func parseBytes(s string) (int64, error) {
 	}
 
 	// Lightbend-faithful per-family fractional handling:
-	// integer-form → int64 * mult; fractional → int64(f) * mult (truncate toward zero).
+	// integer-form → int64 * mult; fractional → int64(f * float64(mult)) (truncate AFTER multiply).
 	if !hasFrac && integerRegex.MatchString(numStr) {
 		n, err := strconv.ParseInt(numStr, 10, 64)
 		if err != nil {
@@ -498,7 +498,7 @@ func parseBytes(s string) (int64, error) {
 	if err != nil {
 		return 0, err
 	}
-	return int64(f) * mult, nil
+	return int64(f * float64(mult)), nil
 }
 
 // ── slices ────────────────────────────────────────────────────────
